requirements.txt
 langchain
 chromadb
 tiktoken
 chainlit
 openai
 langchain-community
 python-dotenv